+++ test/unit/console_tool/test_download_file.py
@@ -54,13 +54,19 @@ def uploaded_file(b2_cli, bucket, local_file):
     }
 
 
-def test_download_file_by_name(b2_cli, local_file, uploaded_file, tmp_path):
+@pytest.mark.parametrize(
+    'flag,expected_stdout', [
+        ('--noProgress', EXPECTED_STDOUT_DOWNLOAD),
+        ('-q', ''),
+        ('--quiet', ''),
+    ]
+)
+def test_download_file_by_name(b2_cli, local_file, uploaded_file, tmp_path, flag, expected_stdout):
     output_path = tmp_path / 'output.txt'
 
     b2_cli.run(
         [
-            'download-file-by-name', '--noProgress', uploaded_file['bucket'],
-            uploaded_file['fileName'],
+            'download-file-by-name', uploaded_file['bucket'], uploaded_file['fileName'],
             str(output_path)
         ],
         expected_stdout=EXPECTED_STDOUT_DOWNLOAD
@@ -68,34 +74,19 @@ def test_download_file_by_name(b2_cli, local_file, uploaded_file, tmp_path):
     assert output_path.read_text() == uploaded_file['content']
 
 
-def test_download_file_by_name_quietly(b2_cli, uploaded_file, tmp_path):
+@pytest.mark.parametrize(
+    'flag,expected_stdout', [
+        ('--noProgress', EXPECTED_STDOUT_DOWNLOAD),
+        ('-q', ''),
+        ('--quiet', ''),
+    ]
+)
+def test_download_file_by_id(b2_cli, uploaded_file, tmp_path, flag, expected_stdout):
     output_path = tmp_path / 'output.txt'
 
     b2_cli.run(
-        [
-            'download-file-by-name', '--quiet', uploaded_file['bucket'], uploaded_file['fileName'],
-            str(output_path)
-        ],
-        expected_stdout=''
+        ['download-file-by-id', flag, '9999', str(output_path)], expected_stdout=expected_stdout
     )
-    assert output_path.read_text() == uploaded_file['content']
-
-
-def test_download_file_by_id(b2_cli, uploaded_file, tmp_path):
-    output_path = tmp_path / 'output.txt'
-
-    b2_cli.run(
-        ['download-file-by-id', '--noProgress', '9999',
-         str(output_path)],
-        expected_stdout=EXPECTED_STDOUT_DOWNLOAD
-    )
-    assert output_path.read_text() == uploaded_file['content']
-
-
-def test_download_file_by_id_quietly(b2_cli, uploaded_file, tmp_path):
-    output_path = tmp_path / 'output.txt'
-
-    b2_cli.run(['download-file-by-id', '--quiet', '9999', str(output_path)], expected_stdout='')
     assert output_path.read_text() == uploaded_file['content']