+++ test/unit/test_console_tool.py
@@ -2173,6 +2173,7 @@ class TestConsoleTool(BaseConsoleToolTest):
         c
         '''
         self._run_command(['ls', '--recursive', 'my-bucket'], expected_stdout, '', 0)
+        self._run_command(['ls', '-r', 'my-bucket'], expected_stdout, '', 0)
 
         # Check long output.   (The format expects full-length file ids, so it causes whitespace here)
         expected_stdout = '''